# [WIP] Performance Gap Between dot & idot