--- a/src/NiSparseArrays.jl
+++ b/src/NiSparseArrays.jl
@@ -1,5 +1,10 @@
 module NiSparseArrays
 
-# Write your package code here.
+using LinearAlgebra: eltype
+using LinearAlgebra
+using NiLang
+using SparseArrays: SparseMatrixCSC, AbstractSparseMatrix, nonzeros, rowvals, nzrange
+
+include("linalg.jl")
 
 end