--- a/src/NiSparseArrays.jl
+++ b/src/NiSparseArrays.jl
@@ -1,5 +1,6 @@
 module NiSparseArrays
 
+using LinearAlgebra: eltype, length
 using LinearAlgebra, SparseArrays
 using NiLang
 using NiLang.AD