--- a/src/sparsegrad.jl
+++ b/src/sparsegrad.jl
@@ -1,13 +1,7 @@
 ###calculate jacobian by NiLang###
-"""
-    helper function params and fitparams
-        function params wraps parameters in track of gradient when given array/matrix.
-            params(v::DenseArray)
-            params(A::SparseMatrixCSC)
-        function fitparams fits the original datastructure with parameters in track of gradient.
-            fitparams(v::DenseArray)
-            fitparams(A::SparseMatrixCSC)
-"""
+
+# function params wraps parameters in track of gradient when given array/matrix.
+# function fitparams fits the original datastructure with parameters in track of gradient.
 params(v::DenseArray{<:Real}) = v
 fitparams(v::DenseArray{<:Real}, p) = p
 params(v::DenseArray{<:Complex{T}}) where T = collect(reinterpret(T, v))
@@ -16,8 +10,9 @@ params(A::SparseMatrixCSC) = params(A.nzval)
 fitparams(A::SparseMatrixCSC, p) = SparseMatrixCSC(A.m, A.n, A.colptr, A.rowval, fitparams(A.nzval, p))
 
 """
-    mvloss(l, j, y::AbstractArray{<:Real}, A, x) -> scalar l 
-mvloss defines loss on sparse-matrix & dense vector/matrix multiplication(Real Type)   
+    mmloss(l, j, y::AbstractArray{<:Real}, A, x) -> scalar l 
+
+mmloss defines loss on sparse-matrix & dense vector/matrix multiplication(Real Type)   
 """
 @i function mmloss(l, j, y::AbstractArray{<:Real}, A, x)
     imul!(y, A, x, 1.0, 1.0)
@@ -25,8 +20,9 @@ mvloss defines loss on sparse-matrix & dense vector/matrix multiplication(Real T
 end
 
 """
-    mvloss(l, j, y::AbstractArray{<:Complex}, A, x) -> scalar l 
-mvloss defines loss on sparse-matrix & dense vector/matrix multiplication(Complex Type)   
+    mmloss(l, j, y::AbstractArray{<:Complex}, A, x) -> scalar l 
+
+mmloss defines loss on sparse-matrix & dense vector/matrix multiplication(Complex Type)   
 """
 @i function mmloss(l, j, y::AbstractArray{<:Complex}, A, x)
     imul!(y, A, x, 1.0, 1.0)
@@ -38,8 +34,9 @@ mvloss defines loss on sparse-matrix & dense vector/matrix multiplication(Comple
 end
 
 """
-    nilang_mv_jacobian(A, x::AbstractArray) -> AbstractArray J
-nilang_mv_jacobian calculates the jacobian with respect to parameter A and x based on mvloss in NiLang
+    nilang_mm_jacobian(A, x::AbstractArray) -> AbstractArray J
+
+nilang_mm_jacobian calculates the jacobian with respect to parameter A and x based on mvloss in NiLang
 """
 function nilang_mm_jacobian(A, x::AbstractArray)
     px, pA = params(x), params(A)