--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,6 +1,8 @@
 using NiSparseArrays
 using Test
+using LinearAlgebra
 
 @testset "NiSparseArrays.jl" begin
-    # Write your tests here.
+    include("linalg.jl") #add sparse multiplication 
+    #include("sparsegrad.jl") #add sparse gradient test
 end