--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -1,6 +1,8 @@
+using LinearAlgebra: include
 using NiSparseArrays
 using Test, Random, LinearAlgebra, NiLang, ForwardDiff, SparseArrays
 using NiLang.AD
 @testset "NiSparseArrays.jl" begin
+    include("utils.jl")
     include("linalg.jl") #add sparse multiplication 
 end