In last PR from dev, I pointed that there may be performance gap between idot and [original implementation](https://github.com/JuliaLang/julia/blob/master/stdlib/SparseArrays/src/linalg.jl#L354) and the order of for-loop may influence the performance.

In my implementation, there triple for-loops while only one for-loop in original implementation since the calculation results of inner double for-loop in my program is equivalent to the _spdot in original implementation (but I couldn't find the source code of _spdot, so there may be some performance tricks here).

I test all the methods in my laptop (macOS intel i5)
Benchmark Performance on  idot(r, x::SparseVector, A::AbstractSparseMatrix{T1}, y::SparseVector{T2}) where {T1, T2}

- Benchmark on SparseArrays dot function:

``` julia
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  251.518 μs … 667.702 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     258.729 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   268.764 μs ±  29.530 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆█▇▇▆▄▄▃▂▁▁ ▁▁▁▁▁▁▁       ▁                                   ▂
  ████████████████████████▇█████▇█▇▇▇▇▇▇███▇▇▇▇▇▆▆▆▆▆▆▇▇▅▆▅▆▅▄▆ █
  252 μs        Histogram: log(frequency) by time        395 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

```

- Benchmark on NiSparseArrays idotxA function:

``` julia
BenchmarkTools.Trial: 2130 samples with 1 evaluation.
 Range (min … max):  2.174 ms …   7.712 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.256 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.327 ms ± 212.084 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▆▄▁
  ████▆▇▆▆▄▅▄▄▄▅▅▄▄▄▃▄▃▄▄▃▃▃▃▃▃▃▃▃▃▃▃▃▃▃▂▂▃▃▂▂▂▂▂▂▂▁▂▂▁▂▂▂▂▂▂ ▃
  2.17 ms         Histogram: frequency by time        2.89 ms <

 Memory estimate: 224 bytes, allocs estimate: 4.

```

- Benchmark on NiSparseArrays idotAx function:

``` julia
BenchmarkTools.Trial: 1308 samples with 1 evaluation.
 Range (min … max):  2.953 ms … 40.413 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.714 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.781 ms ±  1.060 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

               ██▅▃▅▃▅▅▇▄▅▃▃▂▄▄▂▁
  ▃▁▁▁▁▂▁▃▁▂▂▁▁███████████████████▅▇▅▆▄▅▄▃▃▃▂▂▃▂▂▂▂▁▁▂▁▁▂▁▂▂ ▄
  2.95 ms        Histogram: frequency by time        4.89 ms <

 Memory estimate: 224 bytes, allocs estimate: 4.

```

In dev branch, I submitted the idotxA function since I think julia may do some acceleration on it just before I did benchmark tests (for each iteration ,x is always keep the same while A's column varies). However, there is still a relatively big performance gap between idotxA and dot. A more delicate implementation may be considered in the future.