@GiggleLiu 2021-08-29
Since we will revisit all performance issues later, while this benchmark is based on legacy codes, let us close this PR for now.